+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - Fix defend d20Formula code error.
+- Fix defend dialog column formatting.
 - Add roll to check infestation triggers.
 
 # 0.1.0